Fix convert str to float

+++ intelli-analyze/Павлов Алексей Андреевич/lab2.py
@@ -3,7 +3,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 data = pd.read_csv('./flavors_of_cacao.csv', header=0)
-data['Cocoa Percent'] = data['Cocoa Percent'].replace('%', '', regex=True)
+data['Cocoa Percent'] = data['Cocoa Percent'].replace('%', '', regex=True).astype('float64')
 
 
 def possibilities(info, condition):
